requirements.txt
agora==1.2.3
agora_python_server_sdk==2.1.4
agora_token_builder==1.0.0
argh==0.31.3
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
dateutils==0.6.12
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Markdown==3.7
ndg-httpsclient==0.5.1
numpy==2.1.3
onyx.core==1.0.5
packaging==24.2
psycopg2==2.9.10
pusher==3.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyee==12.1.1
PyJWT==2.10.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.14.1
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zope.interface==7.1.1